requirements.txt
Flask==3.0.0
pyyaml==6.0.1
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
pytest==7.4.3
pytest-flask==1.3.0
